requirements.txt
# basic
tqdm
yaml
Pillow
numpy
opencv-python

# load json
# orjson